Fix ruby-prof loading issues
See:

https://github.com/ruby-prof/ruby-prof/pull/300

+++ omnibus/config/software/more-ruby-cleanup.rb
@@ -52,7 +52,6 @@ build do
       diagrams
       example
       examples
-      ext
       Gemfile.lock
       install.rb
       java
@@ -72,6 +71,8 @@ build do
     Dir.glob("#{target_dir}/*/{#{files.join(",")}}").each do |f|
       # chef stores the powershell dlls in the ext dir
       next if File.basename(File.expand_path("..", f)).start_with?("chef-")
+      # ruby-prof has issues/bugs with needing the so in the ext dir
+      next if File.basename(File.expand_path("..", f)).start_with?("ruby-prof-")
 
       puts "Deleting #{f}"
       FileUtils.rm_rf(f)
